Implement blend ops (#49)
* Fix some warning and best practice custom target

* Separate ops into sub-folder and move runtime ops to func calls pass earlier

* Add interfaces for ElementWise Unary/Binary Ops

* remove blur example

* All numbers as FP64 internally with conversion where necessary & common rewriter for element wise binary op with interface implementation for diff & blend ops

* Address reviews from copilot, add lit test

+++ include/ast.h
@@ -54,7 +54,7 @@ struct StringNode : public ASTNode {
  * @brief AST node for numeric literals.
  */
 struct NumberNode : public ASTNode {
-  int64_t value;
+  double value;
   std::string toString() const override;
 };
